requirements.txt
albumentations==1.3.1
anyio==3.7.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.4
colorama==0.4.6
Cython==0.29.36
exceptiongroup==1.1.2
filelock==3.12.2
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
imageio==2.31.1
Jinja2==3.1.2
joblib==1.3.1
lazy_loader==0.3
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.1
numpy==1.25.0
opencv-python==4.8.0.74
opencv-python-headless==4.8.0.74
packaging==23.1
Pillow==10.0.0
python-telegram-bot==20.3
PyWavelets==1.4.1
PyYAML==6.0
qudida==0.0.4
requests==2.31.0
scikit-image==0.21.0
scikit-learn==1.3.0
scipy==1.11.1
sniffio==1.3.0
sympy==1.12
threadpoolctl==3.1.0
tifffile==2023.7.4
torch==2.0.1
torchvision==0.15.2
typing_extensions==4.7.1
urllib3==2.0.3
youtokentome==1.0.6
